refactor: Use dev environment with hot reloading and locked Node version
