+++ package.json
@@ -1,13 +1,15 @@
 {
   "name": "peribolos-as-a-service",
-  "version": "1.0.0",
+  "version": "0.0.1",
+  "private": true,
   "homepage": "https://github.com/open-services-group/peribolos-as-a-service",
   "scripts": {
     "build": "tsc",
-    "start": "probot run ./lib/index.js",
+    "start": "node lib/index.js",
+    "dev": "nodemon src/index.ts",
     "test": "jest",
     "format": "prettier -w 'src/**/*.ts'",
-    "lint": "prettier --check 'src/**/*.ts'; eslint src --ext .ts"
+    "lint": "prettier --check 'src/**/*.{ts,js}'; eslint src --ext .ts,.js"
   },
   "dependencies": {
     "@kubernetes/client-node": "^0.16.3",
@@ -22,9 +24,11 @@
     "eslint-config-prettier": "^8.5.0",
     "jest": "^26.6.3",
     "nock": "^13.0.5",
+    "nodemon": "^2.0.15",
     "prettier": "^2.6.2",
     "smee-client": "^1.0.1",
     "ts-jest": "^26.4.4",
+    "ts-node": "^10.7.0",
     "typescript": "^4.6.3"
   },
   "engines": {
@@ -32,13 +36,26 @@
   },
   "eslintConfig": {
     "root": true,
-    "parser": "@typescript-eslint/parser",
-    "plugins": [
-      "@typescript-eslint"
+    "env": {
+      "node": true,
+      "es6": true
+    },
+    "overrides": [
+      {
+        "files": [
+          "**/*.ts"
+        ],
+        "parser": "@typescript-eslint/parser",
+        "plugins": [
+          "@typescript-eslint"
+        ],
+        "extends": [
+          "plugin:@typescript-eslint/recommended"
+        ]
+      }
     ],
     "extends": [
       "eslint:recommended",
-      "plugin:@typescript-eslint/recommended",
       "prettier"
     ]
   },
